Code refactor. Beta loader for modules, added test module

+++ unix-api/src/main/kotlin/net/unix/api/modification/ModificationInfo.kt
@@ -15,7 +15,7 @@ interface ModificationInfo : Serializable {
     /**
      * Load priority. The higher the number, the higher the priority
      */
-    val priority: UInt
+    val priority: Double
 
     /**
      * Modification name.
